feat(templating): Update rearrange and update readme
